Product-Management/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sqlalchemy==2.0.29
pymysql==1.1.0
pydantic-settings==2.2.1
python-dotenv==1.0.1